BUG: 116831
This fixes a few layout problems in RTL desktops. The code is fully documented
whith the changes I made, but be aware - they are ugly. But, hell... this works :)

Todo for Qt4:
If someone wants up port the fixes to Qt4, please don't. Qt4 has better ways of doying
this, (see the direction property of each widget). 



svn path=/branches/KDE/3.5/kdeedu/kbruch/; revision=503310