backport from Trunk (https://bugs.kde.org/token.cgi?t=8cDojiBlcS&a=cfmem):
Implements a better way to check if a task was solved correctly by using
the == operator of the ratio class itself (fix bug 178007).

In contrast to Trunk, I have not added a help message to explain users why
tasks are counted as wrong if they were not entered in mixed notation. This
is not possible due to string freeze!

additional changes:
- removed a lot of debug statements from the code
- added clean() function to task; useful for debugging
- some whitespace changes introduced by kdevelop

Changes are partly based on patch provided by Tadeu Araujo. Thanks for your
contribution!

BUG:178007

svn path=/branches/KDE/4.2/kdeedu/kbruch/; revision=963048