Fix call to QStringLiteral, pass the proper path
The function was using also other strings in addition to the intended
one, leading to an incorrect stylesheet and bigger text.

Discovered thanks to the warning printed on the command line:
"Could not parse stylesheet of object 0x..."

(Side note: the text size should not be pixel-based, but this is
a bigger change).