Add proper spacing around operators and equals sign
In proper mathematical typography, there should be a space
on each side of arithmetic operators and the equals sign.

Note that a space have *not* been added around the equals
sign string in the factorize exercise, as this equal sign
is painted using QGridLayout, which already adds some spacing
between each cell.